+++ CHANGELOG.md
@@ -1,8 +1,38 @@
+# Version 0.21.0 (2023-09-29)
+
+## API
+
+- Add `reader-mp4` feature
+- Drop MPEG-1/2 Audio Layer 1 (MP1) support
+
+## Fixes
+
+- Fix MPEG-DASH Manifest (MPD) name
+
+## Improvements
+
+- Add precision to the 3rd Generation Partnership Project (3GPP) signature
+- Add precision to the 3rd Generation Partnership Project 2 (3GPP2) signature
+- Add precision to the High Efficiency Image File Format (HEIF) signature
+- Add precision to the High Efficiency Image File Format Sequence (HEIFS) signature
+- Add precision to the JPEG 2000 Part 1 (JP2) signature
+- Add precision to the MPEG-4 Part 14 (MP4) signature
+
+## New formats support
+
+- ActiveMime (MSO)
+- Amiga Disk File (ADF)
+- MPEG-4 Part 14 (MP4)
+- MPEG-4 Part 14 Audio (MP4)
+- MPEG-4 Part 14 Subtitles (MP4)
+- Matroska Subtitles (MKS)
+- Silicon Graphics Movie (SGI)
+
 # Version 0.20.0 (2023-09-07)
 
 ## New formats support
 
-- MPEG-DASH manifest (MPD)
+- MPEG-DASH Manifest (MPD)
 - Microsoft Works 6 Spreadsheet (XLR)
 - Microsoft Works Database (WDB)
 - Microsoft Works Spreadsheet (WKS)